% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDag.R
\name{buildGraph}
\alias{buildGraph}
\title{Builds a dagitty object saturated (fully connected) graph}
\usage{
buildGraph(
  type = c("full", "saturated", "ordered"),
  variables = NULL,
  treatment = NULL,
  outcome = NULL,
  confounders = NULL,
  mediators = NULL,
  latent_variables = NULL,
  instrumental_variables = NULL,
  mediator_outcome_confounders = NA,
  coords_spec = c(lambda = 0, lambda_max = NA, iterations = NA)
)
}
\arguments{
\item{type}{The type of graph generated. Defaults to 'full', producing a fully connected graph with confounders connected in both directions (bi-directional), and to mediators in one direction (uni-directional). If type ='saturated', a similar saturated graph is produced except confounders are not connected to mediators, featuring bi-directional arrows between each of the confounders (follows the ESC-DAGs Mapping Stage in Ferguson et al. (2020)). When type = 'ordered', the order of supplied confounders and mediators determines the order that each node occurs, therefore directed arrows are to be connected in one direction from confounders and mediators to other confounders and mediators, respectively. This builds a saturated DAG with temporal, uni-directional arrows, based on Tennnant et al. (2021).}

\item{variables}{Vector of variables to be assigned nodes in a graph.  A named vector can be supplied, containing any of the other input variable types}

\item{treatment}{Treatment variable, e.g. "X". Must be specified, unless included in the named vector 'variables'.}

\item{outcome}{Outcome variable, e.g. "Y". Must be specified, unless included in the named vector 'variables'.}

\item{confounders}{Vector of confounder variables, e.g. c("Z1", "Z2", "Z3").}

\item{mediators}{Vector of mediator variables, e.g. c("M1", "M2", "M3").}

\item{latent_variables}{Vector of already supplied variable names to be labelled latent (unobserved), e.g. c("L1", "L2", "M1").}

\item{instrumental_variables}{Vector of instrumental variables, e.g. "IV"}

\item{mediator_outcome_confounders}{Vector of already supplied confounder names, that instead of being common causes of treatment and outcome (X <- Z -> Y) are a common cause of mediators and outcome (M <- Z -> Y). A list of vectors can instead be supplied with separate vectors for each confounder name, and their connected mediator(s).}

\item{coords_spec}{Set of parameters for generating coordinates. Adjust node placement with lambda, a higher value increases volatility and results in more extreme DAG structures. Setting 'lambda_max' generates a DAG at each lambda value between lambda and lambda_max (only used if iterations is supplied). Iterations controls number of repeats for each lambda value (returns the first lambda value if NULL).}
}
\value{
A dagitty object, fully connected (saturated) graph.
}
\description{
buildGraph() produces a saturated graph by default from the supplied treatment, outcome, confounder, and  mediators (optional) based on the type of graph specified. Optional latent confounder or medfiator variables can be specified.
}
\examples{


# There are three main ways to supply variables to buildGraph().

# Option 1: Inputting a named vector of variables inputted (can either include treatment and outcome or as separate inputs)
{
  variables <- c(confounders = c("Z1", "Z2", "Z3"),
                 treatment = "X",
                 outcome = "Y",
                 mediators = "M",
                 instrumental_variables = "IV")
}

# Three graph types can also be generated. First, we build an 'ordered' graph where the supplied vector order is used to determine the temporal order of confounder nodes.
  type <- "ordered"

  dag <- buildGraph(type = type,
                    variables = variables)

# Plotting the graph using ggdagitty() assigns coordinates based on the variable roles (confounder, treatment, outcome, mediator, etc.).
ggdagitty(dag)

#' # Option 2: The 'confounders' input can be used, while 'variables' is ignored.
#           Separate inputs are required for treatment, outcome, and other nodes.
{
  confounders <- c("Z1", "Z2", "Z3")
  treatment <- "X"
  outcome <- "Y"
  instrumental_variables <- "IV"
  mediator = "M"
}

A "saturated" graph type bidirectionally connects each of the confounders, in addition to directed arrows to the outcome and treatment nodes.
type <- "saturated"

dag <- buildGraph(type = type,
                  treatment = treatment,
                  outcome = outcome,
                  confounders = confounders,
                  mediators = mediators,
                  instrumental_variable = instrumental_variable)

ggdagitty(dag, labels)

# Option 3: The 'variables' input can be used to connect all variables to each other, with treatment, outcome and other nodes inputted separately.
#           Inputted variables are treated as confounders, while the actual 'confounders' input can be left blank.
#           We decided to keep this option, rather than remove it, to provide easier access to users who may not have much 'exposure' to causal graphs and terminology.
{
  variables <- c("Z1", "Z2", "Z3")
  treatment <- "X"
  outcome <- "Y"
  instrumental_variables <- "IV"
  mediator = "M"
}

# Using type = "full" generates a fully connected graph: all confounders are connected to each other, in both directions, and also to mediators, treatment and outcome.
type <- "full"

dag <- buildGraph(type = type,
                  variables = variables,
                  treatment = treatment,
                  outcome = outcome,
                  mediators = mediators,
                  instrumental_variables = instrumental_variables)

ggdagitty(dag)

}
