% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{causaliflower-package}
\alias{causaliflower}
\alias{causaliflower-package}
\title{causaliflower: Build and Assess Causal Graphs}
\description{
The causaliflower package aims to provide simple ways of working with causal graphs in R. It extends 'dagitty' and 'ggdag' functions to enable cohesive and reproducible causal analysis workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AAGI-AUS/causaliflower}
  \item Report bugs at \url{https://github.com/AAGI-AUS/causaliflower/issues}
}

}
\author{
\strong{Maintainer}: Aidan Moller \email{aidan.moller@adelaide.edu.au} [copyright holder]

}
\keyword{internal}
