% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_models.r
\name{causalify_glm}
\alias{causalify_glm}
\title{GLM with dagitty}
\usage{
causalify_glm(dag, data, family = gaussion())
}
\arguments{
\item{dag}{dagitty object}

\item{data}{A data frame or data table, with variables corresponding to nodes in the supplied dag.}

\item{family}{Used for glm and gamm models, defaults to gaussion().}

\item{model}{Current options for model include "lm", "glm", "lme", or "gamm". The last two options are taken from the 'mgcv' package, with "lme" returning the linear mixed-effect model output from mgcv::gamm().}
}
\value{
Model object of either
}
\description{
causalify_glm()
}
