% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalify_models.r
\name{causalify_lm}
\alias{causalify_lm}
\title{DAG Linear Model}
\usage{
causalify_lm(dag, data)
}
\arguments{
\item{dag}{dagitty object}

\item{data}{A data frame or data table, with variables corresponding to nodes in the supplied dag.}
}
\value{
Model object of either
}
\description{
causalify_lm()
}
