% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_models.r
\name{causalify_lme}
\alias{causalify_lme}
\title{GAMM with dagitty}
\usage{
causalify_lme(
  dag,
  data,
  random_effects = NULL,
  method = "REML",
  family = gaussion(),
  corr_structure = "corAR1"
)
}
\arguments{
\item{dag}{dagitty object}

\item{data}{A data frame or data table, with variables corresponding to nodes in the supplied dag.}

\item{random_effects}{A vector or list of variables to be treated as random effects. Only used when "gamm" or "gam" is specified, the  are not required to be included in the dag.}

\item{method}{Defaults to method = "REML", only used for gamm models.}

\item{family}{Used for glm and gamm models, defaults to gaussion().}

\item{corr_structure}{Limited functionality. Currently only used when model = "lme" or "gamm" to specify nlme::corAR1(0, form=~1|year) if a random_effects input contains the string 'year'.}
}
\value{
Model object of either
}
\description{
causalify_gamm()
}
