% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{colliders}
\alias{colliders}
\title{Collider names in a dagitty object}
\usage{
colliders(dag)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
Character vector of collider names; character(0) if none.
}
\description{
colliders() identifies collider nodes that are common children of both treatment and outcome (treatment -> C <- outcome) in a directed acyclic graph; it does not return arbitrary two-parent colliders.
}
\examples{
set.seed(1)
dag <- build_graph(
  variables = c("Z1", "Z2"), treatments = "X", outcomes = "Y",
  mediators = "M", instrumental_variables = "IV", type = "saturated"
)
colliders(dag)

}
