% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{competing_causes}
\alias{competing_causes}
\title{Competing cause node names in a dagitty object}
\usage{
competing_causes(dag)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
Character vector of competing cause names; character(0) if none.
}
\description{
competing_causes() returns direct causes of the outcome that are not otherwise classified -- i.e. causes of the outcome lying off the treatment->outcome path and unconnected to the treatment. Sometimes called a competing exposure (Tennant et al., 2021).
}
\examples{
set.seed(1)
dag <- build_graph(
  variables = c("Z1", "Z2"), treatments = "X", outcomes = "Y",
  mediators = "M", instrumental_variables = "IV", type = "saturated"
)
competing_causes(dag)

}
