% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{competing_exposures}
\alias{competing_exposures}
\title{Renamed: competing_exposures}
\usage{
competing_exposures(dag)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
No return value; calling this function stops with a renaming message.
}
\description{
competing_exposures() was renamed competing_causes(). Calling this name stops
with a pointer to the new one; it will be removed in a later release.
}
