% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{confounders}
\alias{confounders}
\title{Confounder names in a dagitty object}
\usage{
confounders(dag)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
Character vector of confounder names; character(0) if none.
}
\description{
confounders() identifies common causes of treatment and outcome in a directed acyclic graph. Ancestral common causes are included, as are proxy variables of indicated unmeasured confounding.
}
\examples{
set.seed(1)
dag <- build_graph(
  variables = c("Z1", "Z2"), treatments = "X", outcomes = "Y",
  mediators = "M", instrumental_variables = "IV", type = "saturated"
)
confounders(dag)

}
