% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_funcs.R
\name{copy_nodes}
\alias{copy_nodes}
\title{Replicate nodes in a dagitty object}
\usage{
copy_nodes(
  dag,
  existing_nodes,
  existing_node_type = "pre_treatment",
  new_node_type = "post_treatment",
  temporal_reference_node = NA,
  num_repeats = NA,
  coords_spec = c(lambda = 0.1, threshold = 0.5)
)
}
\arguments{
\item{dag}{A dagitty object. Must include exposure and outcome nodes.}

\item{existing_nodes}{Vector of existing node names, used as a reference for the new graph nodes, e.g., c("Z1", "Z2", "Z3").}

\item{existing_node_type}{A suffix added to each of the reference node names, e.g. "pre_treatment", or "t0".}

\item{new_node_type}{A suffix added to each of the new node names, e.g. "post_treatment", or "t" (a number is added for each repeat if num_repeats is specified)}

\item{temporal_reference_node}{Supply an alternative reference, or simply leave blank. Default settings use treatment as the temporal point of reference for adding new (post-treatment) nodes, but if other node types are specified it can be useful to specify the existing node name.}

\item{num_repeats}{Number of additional copies of nodes, such as time points. Each repeat number is included at the end of new node names (new_new_t1, new_node_t2, etc.).}

\item{coords_spec}{Set of parameters for generating coordinates. Adjust node placement with lambda, a higher value increases volatility and results in more extreme DAG structures.}
}
\value{
A dagitty object.
}
\description{
Create single or multiple copy nodes based on existing nodes, or another dagitty object.
}
\examples{
dag <- copy_nodes(dag, existing_nodes)

}
