% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.r
\name{create_new_node_names}
\alias{create_new_node_names}
\title{Add nodes to a dagitty object}
\usage{
create_new_node_names(reference_nodes, new_node_type, num_repeats)
}
\arguments{
\item{reference_nodes}{Vector of existing node names, used as a reference for the new graph nodes, e.g., c("Z1", "Z2", "Z3").}

\item{new_node_type}{A suffix added to each of the new node names, e.g. "post_treatment", or "t" (a number is added for each repeat if num_repeats is specified)}

\item{num_repeats}{Number of additional copies of nodes, such as time points. Each repeat number is included at the end of new node names (new_new_t1, new_node_t2, etc.).}
}
\value{
A vector of new node names.
}
\description{
create_new_node_names() is a helper function for addNodes(). It uses existing nodes to create new node names for specified repeats/time points.
}
