% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funcs.r
\name{get_diff_edges}
\alias{get_diff_edges}
\title{Compare dagitty object edges}
\usage{
get_diff_edges(dag1, dag2, compare_all = TRUE, include_roles = FALSE)
}
\arguments{
\item{dag1}{A dagitty object.}

\item{dag2}{A dagitty object.}

\item{compare_all}{Return mutually exclusive edges in both dag1 and dag2 (compare_all = TRUE) e.g. dag1 edges not in dag2 AND dag2 edges not in dag1, or only dag1 mutually exclusive edges (compare_all = FALSE), e.g. dag1 edges not in dag2. Defaults to compare_all = TRUE.}

\item{include_roles}{Determines whether to include node roles in addition to node names and edge direction. Defaults to include_roles = FALSE, e.g. only considers edges where descendant/ancestor node names and edge direction are different.}
}
\value{
Data table of edges.
}
\description{
Compare dagitty object edges
}
\examples{
set.seed(1)
dag <- build_graph(
  variables = c("Z1", "Z2"), treatments = "X", outcomes = "Y", type = "saturated"
)
dag2 <- build_graph(
  variables = c("Z1", "Z3"), treatments = "X", outcomes = "Y", type = "saturated"
)
get_diff_edges(dag, dag2)

}
