% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funcs.r
\name{get_diff_roles}
\alias{get_diff_roles}
\title{Compare dagitty object node roles}
\usage{
get_diff_roles(dag1, dag2, compare_all = TRUE, nodes_in_common = FALSE)
}
\arguments{
\item{dag1}{A dagitty object.}

\item{dag2}{A dagitty object.}

\item{compare_all}{Compare and return mutually all exclusive roles  (compare_all = TRUE) e.g. dag1 roles not in dag2 AND dag2 roles not in dag1, or set compare_all = FALSE to return only dag1 mutually exclusive roles, e.g. dag1 roles not in dag2. Defaults to compare_all = TRUE.}

\item{nodes_in_common}{Determines whether to include only the nodes found in both DAGs. Defaults to nodes_in_common = FALSE (returns all nodes).}
}
\value{
A list of nodes.
}
\description{
Compare dagitty object node roles
}
\examples{
set.seed(1)
dag <- build_graph(
  variables = c("Z1", "Z2"), treatments = "X", outcomes = "Y", type = "saturated"
)
dag2 <- build_graph(
  variables = c("Z1", "Z3"), treatments = "X", outcomes = "Y", type = "saturated"
)
get_diff_roles(dag, dag2)

}
