% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funcs.r
\name{get_nodes}
\alias{get_nodes}
\title{dagitty node names and roles}
\usage{
get_nodes(dag)
}
\arguments{
\item{dag}{dagitty object}
}
\value{
Nested list of nodes and node relationships
}
\description{
get_nodes() is a dagitty wrapper that returns a list of node names and their roles.
}
\examples{
set.seed(1)
dag <- build_graph(
  variables = c("Z1", "Z2"), treatments = "X", outcomes = "Y",
  mediators = "M", instrumental_variables = "IV", type = "saturated"
)
get_nodes(dag)

}
