% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funcs.r
\name{get_nodes_from_treatment_to_outcome}
\alias{get_nodes_from_treatment_to_outcome}
\title{nodes in path from treatment to outcome}
\usage{
get_nodes_from_treatment_to_outcome(dag, output_list = FALSE, directed = FALSE)
}
\arguments{
\item{dag}{A dagitty object}

\item{output_list}{TRUE or FALSE to output a list (default FALSE returns a vector).}
}
\value{
Vector or list of  nodes in the path from treatment to outcome.
}
\description{
get_nodes_from_treatment_to_outcome() is a dagitty and ggdag::dag_paths() wrapper function, intended for use with multiple treatments and outcomes.
}
