% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdagitty.R
\name{getCoords}
\alias{getCoords}
\title{Add coordinates to a dagitty object}
\usage{
getCoords(
  dag,
  coords_spec = c(lambda = 0, lambda_max = NA, iterations = NA),
  confounders = NULL
)
}
\arguments{
\item{dag}{Dagitty object, with at least a treatment and outcome set. Nodes are automatically placed in the following categories: treatment, outcome, confounder, mediator, latent variable, mediator-outcome-confounder, or instrumental variable.}

\item{coords_spec}{Set of parameters for generating coordinates. Adjust node placement with lambda, a higher value increases volatility and results in more extreme DAG structures. Setting 'lambda_max' generates a DAG at each lambda value between lambda and lambda_max (only used if iterations is supplied). Iterations controls number of repeats for each lambda value (returns the first lambda value if NULL).}

\item{confounders}{Vector of confounders in order of occurrence. If left blank, the order is inferred with dagitty::topologicalOrdering().}
}
\value{
dagitty objecty with coordinates.
}
\description{
Add coordinates to a dagitty object
}
