% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDag.R
\name{getEdges}
\alias{getEdges}
\title{Gets edges between nodes and their roles}
\usage{
getEdges(dag, selected_nodes = NULL, variables = NULL, wide = FALSE)
}
\arguments{
\item{dag}{A dagitty object.}

\item{selected_nodes}{Nodes to return edges. Defaults to NULL, or can be a character or vector combination of any of the following: c("treatment", "outcome", "confounder", "mediator", "latent", "moc", "instrumental")}
}
\value{
A dataframe containing the specified node edges.
}
\description{
getEdges() filters a dagitty object and returns a data frame with edges for specified node roles.
}
