% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDag.R
\name{getFeatureMap}
\alias{getFeatureMap}
\title{Add coordinates to a dagitty object}
\usage{
getFeatureMap(dag)
}
\arguments{
\item{dag}{dagitty object}

\item{confounders}{vector of confounders nodes in the supplied dag.}

\item{mediators}{vector of mediator nodes in the supplied dag.}

\item{instrumental_vars}{vector of instrumental variable nodes in the supplied dag.}

\item{lambda}{adjusts sensitivity of node placement. Higher lambda introduces more volatility in confounder nodes and treatment along the y-axis.}
}
\value{
dagitty objecty with coordinates.
}
\description{
Add coordinates to a dagitty object
}
