% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdagitty.R
\name{getLabels}
\alias{getLabels}
\title{Add labels to a dagitty object
Extract node intials to use as labels, essentially a wrapper function for ggdag::dag_label().}
\usage{
getLabels(dag, type = "initials")
}
\arguments{
\item{dag}{dagittyy object}
}
\value{
Labelled vector of node names from the dagitty object
}
\description{
Add labels to a dagitty object
Extract node intials to use as labels, essentially a wrapper function for ggdag::dag_label().
}
