% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNDVI.R
\name{getNDVI}
\alias{getNDVI}
\title{Provides NDVI at different time points using satellite data (Sentinel-2)}
\usage{
getNDVI(bbox, start_date, end_date, pixel_x_size, pixel_y_size)
}
\arguments{
\item{bbox}{A 'bbox' object indicating the area to be observed. This can be created using sf::st_bbox() with the argument crs = sf::st_crs(4326).}

\item{start_date}{Start of the date range, "YYYY-MM-DD".}

\item{end_date}{Final date to be observed, "YYYY-MM-DD".}

\item{pixel_x_size}{Longitude resolution using decimal degree notation.}

\item{pixel_y_size}{Latitude resolution using decimal degree notation.}
}
\value{
A dataframe containing columns for NDVI, and observation dates. NDVI is calculated from the mean B04 and B08 asset values at each time point, taken from pixels in the area defined by the 'bbox' object. Mean NDVI is used for duplicate dates.
}
\description{
getNDVI() calculates mean NDVI in an area at a user-specified resolution for Sentinel-2 satellite image data in the supplied range. This function is essentially a wrapper for the 'rsi' functions rsi_query_api() and get_stac_data() that calculates NDVI from Sentinel-2 observations.
}
\examples{
st_bbox <- sf::st_bbox(c(xmin = 150, xmax = 150.5, ymin = -30, ymax = -30.2), crs = sf::st_crs(4326))
start_date <- "2023-01-01"
end_date <- "2023-01-10"
pixel_y_size <- 0.01
pixel_x_size <- 0.01

ndvi_df <- getNDVI(st_bbox, start_date, end_date, pixel_x_size, pixel_y_size)
ndvi_df

}
