% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNDVI.R
\name{getNDVIraster}
\alias{getNDVIraster}
\title{Produces an NDVI image using satellite data (Sentinel-2)}
\usage{
getNDVIraster(bbox, start_date, end_date, pixel_x_size, pixel_y_size)
}
\arguments{
\item{bbox}{A 'bbox' object indicating the area to be observed. This can be created using sf::st_bbox() with the argument crs = sf::st_crs(4326).}

\item{start_date}{Start of the date range, "YYYY-MM-DD".}

\item{end_date}{Final date to be observed, "YYYY-MM-DD".}

\item{pixel_x_size}{Longitude resolution using decimal degree notation.}

\item{pixel_y_size}{Latitude resolution using decimal degree notation.}
}
\value{
A SpatRaster object of NDVI values, calculated from band 4 and band 8 assets. A median composite function is used in rsi::get_stac_date() to combine STAC images, collected from sentinel 2 satellite observation data using rsi::rsi_query_api().
}
\description{
getNDVIraster() produces a raster image of the median NDVI values for the supplied date range, area and resolution. This function is essentially a wrapper for the 'rsi' functions and , which collect and transform satellite image data. Sentinel-2 STAC are queried using rsi_query_api(), and a composite function from rsi::get_stac_data() combines asset layers by taking the median value of pixels for each image collected. The median NDVI is calculated for each pixel in the supplied area and resolution.
}
\examples{
st_bbox <- sf::st_bbox(c(xmin = 150, xmax = 150.5, ymin = -30, ymax = -30.2), crs = sf::st_crs(4326))
start_date <- "2023-01-01"
end_date <- "2023-01-10"
pixel_y_size <- 0.01
pixel_x_size <- 0.01

ndvi_rast <- getNDVIraster(st_bbox, start_date, end_date, pixel_x_size, pixel_y_size)
terra::plot(ndvi_rast)

}
