% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdagitty.R
\name{ggdagitty}
\alias{ggdagitty}
\title{ggdag plot from a dagitty object}
\usage{
ggdagitty(
  dag,
  coords_spec = c(lambda = 0, lambda_max = NA, iterations = NA),
  labels = NULL,
  label_type = "label"
)
}
\arguments{
\item{dag}{dagitty object}

\item{coords_spec}{Set of parameters for generating coordinates. Adjust node placement with lambda, a higher value increases volatility and results in more extreme DAG structures. Setting 'lambda_max' generates a DAG at each lambda value between lambda and lambda_max (only used if iterations is supplied). Iterations controls number of repeats for each lambda value (returns the first lambda value if NULL).}

\item{labels}{vector of labels for nodes in dagitty object}
}
\value{
ggdag plot
}
\description{
Generates a ggdag plot from a dagitty object, adding coordinates and node labels using initials (capital letters).
Essentially a wrapper for dagitty and ggdag plotting functions with added features.
}
