% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_funcs.R
\name{join_graphs}
\alias{join_graphs}
\title{Merge two dagitty objects}
\usage{
join_graphs(dag, new_dag)
}
\arguments{
\item{dag}{First dagitty object.}

\item{new_dag}{Second dagitty object, added to the first.}
}
\value{
A dagitty object
}
\description{
join_graphs() adds new nodes to the first from the second supplied dagitty object, before generating new coordinates for the merged graph.
}
\examples{
set.seed(1)
dag <- build_graph(
  variables = c("Z1", "Z2"), treatments = "X", outcomes = "Y", type = "saturated"
)
dag2 <- build_graph(
  variables = c("Z1", "Z3"), treatments = "X", outcomes = "Y", type = "saturated"
)
new_graph <- join_graphs(dag, dag2)

}
