% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assess_keep_edges.r
\name{keepEdges}
\alias{keepEdges}
\title{Remove dagitty object edges}
\usage{
keepEdges(dag, edges_to_keep = NA)
}
\arguments{
\item{dag}{A saturated graph dagitty object. Exposure and outcome must be indicated, and optionally can include assigned coordinates.}

\item{edges_to_keep}{A vector of directed arrows to be kept between (non-treatment and non-outcome) variables, e.g. c("Z1 -> Z2", "Z2 -> Z3"), c("y", "n", "y"), or c(TRUE, FALSE, TRUE).}
}
\value{
A dagitty object, with directed arrows removed based on edges_to_keep.
}
\description{
keepEdges() removes edges based on user inputs.
}
