% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landsat_index_list}
\alias{landsat_index_list}
\title{Landsat spectral indices list}
\format{
A named list where each element is a list with the following elements:
\describe{
\item{assets}{Character vector of Landast band names.}
\item{fun}{A function that takes a SpatRaster and returns a single-layer SpatRaster.}
\item{res}{Recommended resolution (highest integer value used by assets).}
}
}
\usage{
landsat_index_list
}
\description{
A named list of spectral indices compatible with Landsat 5 & 8 sources, for use with \code{\link{get_rs_data}} and \code{\link{get_rs_raster}}.
}
\keyword{datasets}
