% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_utility_funcs.r
\name{markov_graph}
\alias{markov_graph}
\title{Remove edges using markov logic}
\usage{
markov_graph(dag)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
A dagitty object.
}
\description{
markov_graph()
}
