% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{mediators}
\alias{mediators}
\title{Mediator names in a dagitty object}
\usage{
mediators(dag)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
Character vector of mediator names; character(0) if none.
}
\description{
mediators() identifies nodes along directed paths between treatment and outcome in a directed acyclic graph.
}
\examples{
set.seed(1)
dag <- build_graph(
  variables = c("Z1", "Z2"), treatments = "X", outcomes = "Y",
  mediators = "M", instrumental_variables = "IV", type = "saturated"
)
mediators(dag)

}
