% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph.R
\name{merge_graphs}
\alias{merge_graphs}
\title{Merge two dagitty objects}
\usage{
merge_graphs(dag, new_dag, coords_spec = c(lambda = 1, threshold = 1))
}
\arguments{
\item{dag}{First dagitty object.}

\item{new_dag}{Second dagitty object, added to the first.}

\item{coords_spec}{Parameters used for generating coordinates. Adjust node placement with lambda; a higher value increases volatility and results in more extreme DAG structures. Threshold controls the closeness of nodes.}
}
\value{
A dagitty object
}
\description{
merge_graphs() adds new nodes to the first from the second supplied dagitty object, before generating new coordinates for the merged graph.
}
\examples{
merged_dag <- merge_graphs(dag, new_dag)

}
