% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{minimal_sets}
\alias{minimal_sets}
\title{Minimal sufficient adjustment sets}
\usage{
minimal_sets(
  dag,
  treatment = NULL,
  outcome = NULL,
  effect = "total",
  num_sets = 5
)
}
\arguments{
\item{dag}{A dagitty object.}

\item{treatment}{Vector of treatment(s).}

\item{outcome}{Vector of outcome(s).}

\item{effect}{Defaults to total effect, options available as per dagitty::adjustmentSets()}

\item{num_sets}{Number of sets to return, defaults to the smallest five minimally sufficient adjustment sets.}
}
\value{
Named list of minimally sufficient adjustment sets.
}
\description{
minimal_sets() is a dagitty::adjustmentSets() wrapper for obtaining minimal adjustment sets, returning the smallest 5 (if available) by default.
}
\examples{
set.seed(1)
dag <- build_graph(
  variables = c("Z1", "Z2"), treatments = "X", outcomes = "Y",
  mediators = "M", instrumental_variables = "IV", type = "saturated"
)
minimal_sets(dag) # defaults to the total effect and 5 smallest sets

minimal_sets(dag, effect = "direct") # direct effect

minimal_sets(dag, effect = "direct", num_sets = 1) # return only the smallest set (direct effect)

}
