% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdagitty.R
\name{node_coords}
\alias{node_coords}
\title{Create coordinates for new nodes}
\usage{
node_coords(
  dag,
  new_node_names = NULL,
  coordinates = NA,
  coords_spec = c(lambda = 1, threshold = 1)
)
}
\arguments{
\item{dag}{A dagitty object. Must include exposure and outcome nodes.}

\item{new_node_names}{Vector of new node names. When new nodes are specified, existing nodes can be used as a reference point. Otherwise, all new coordinates are generated using ggdag::automatically generates}

\item{coordinates}{list of coordinates from a dagitty object.}

\item{coords_spec}{Parameters used for generating coordinates. Adjust node placement with lambda; a higher value increases volatility and results in more extreme DAG structures. Threshold controls the closeness of nodes.}
}
\value{
Dagitty coordinates.
}
\description{
Create coordinates for new nodes
}
\examples{
coordinates <- node_coords(dag, coords_spec = c(0.1, 0.5)) # generate new coordinates

dagitty::coordinates(dag) <- coordinates # add coordinates to dagitty object

}
