% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_utility_funcs.r
\name{node_roles}
\alias{node_roles}
\title{dagitty node names and roles}
\usage{
node_roles(dag)
}
\arguments{
\item{dag}{dagitty object}
}
\value{
Nested list of nodes and node relationships
}
\description{
node_roles() is a dagitty wrapper that returns a list of node names and their roles.
}
\examples{
node_roles(dag)

}
