% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.r
\name{nodes_between_treatment_and_outcome}
\alias{nodes_between_treatment_and_outcome}
\title{node names in path from treatment to outcome}
\usage{
nodes_between_treatment_and_outcome(
  dag,
  treatments,
  outcomes,
  output_list = FALSE
)
}
\arguments{
\item{dag}{A dagitty object}

\item{treatments}{A vector of treatment node names.}

\item{outcomes}{A vector of outcome node names.}

\item{output_list}{TRUE or FALSE to output a list (default FALSE returns a vector).}
}
\value{
Vector or list of  nodes in the path from treatment to outcome.
}
\description{
nodes_between_treatment_and_outcome() is a ggdag::dag_paths() wrapper intended for use with multiple treatments and outcomes.
}
