% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_sensing_funcs.R
\name{point_to_bbox}
\alias{point_to_bbox}
\title{Bounding box around a central point}
\usage{
point_to_bbox(longitude, latitude, x_metres, y_metres)
}
\description{
Bounding box around a central point
}
