% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{proxies}
\alias{proxies}
\title{Proxy node names in a dagitty object}
\usage{
proxies(dag)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
Character vector of proxy variable names; character(0) if none.
}
\description{
proxies() is a dagitty wrapper that identifies nodes that are proxy variables for indicated unmeasured confounding (direct children of a latent variable that are direct parents of treatment or outcome) in a directed acyclic graph.
}
\examples{
set.seed(1)
dag <- build_graph(
  variables = c("Z1", "Z2"), treatments = "X", outcomes = "Y",
  mediators = "M", instrumental_variables = "IV", type = "saturated"
)
proxies(dag)

}
