% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_funcs.R
\name{saturate_nodes}
\alias{saturate_nodes}
\title{Connect nodes in a dagitty object}
\usage{
saturate_nodes(
  dag,
  nodes = NULL,
  node_role = NULL,
  type = "full",
  print_edges = TRUE
)
}
\arguments{
\item{dag}{A dagitty object. Must include exposure and outcome nodes.}

\item{nodes}{Inputted vector of node names to be added to the graph.}

\item{node_role}{Determines which nodes will be connected, default fully connects new node to all existing nodes. Other options include the following: c("confounder", "treatment", "outcome", "mediator", "mediator_outcome_confounder", "instrumental", "competing_exposure", "collider", "latent", "observed").}

\item{type}{Type of graph generated. Defaults to 'full' (fully connected graph) with arrows drawn between confounders (both directions) and from confounders to mediators. If type ='saturated', a similar saturated graph is produced except confounders are not connected to mediators, featuring bi-directional arrows between each of the confounders (follows the ESC-DAGs Mapping Stage in Ferguson et al. (2020)). When type = 'first' or 'last', inputted nodes are ordered first or last if a confounder, mediator, or treatment node_role is selected.}

\item{print_edges}{Print new edges in console, defaults to TRUE.}
}
\value{
A dagitty object.
}
\description{
Add nodes to a dagitty object, connecting edges based on the 'type' of graph selected, and generate new node coordinates using existing nodes.
}
\examples{
dag <- saturate_nodes(dag, nodes)

}
