% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{treatments}
\alias{treatments}
\title{Treatment variables in a dagitty object}
\usage{
treatments(dag)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
Character vector of treatment names; character(0) if none.
}
\description{
treatments() returns a vector of node names assigned "\code{[exposure]}" in a dagitty object (similar to dagitty::exposures())
}
\examples{
set.seed(1)
dag <- build_graph(
  variables = c("Z1", "Z2"), treatments = "X", outcomes = "Y",
  mediators = "M", instrumental_variables = "IV", type = "saturated"
)
treatments(dag)

}
