% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDag.R
\name{trimGraph}
\alias{trimGraph}
\title{Trim edges in a graph}
\usage{
trimGraph(dag, edges_to_keep = NA)
}
\arguments{
\item{dag}{A saturated graph dagitty object. Exposure and outcome must be indicated, and optionally can include assigned coordinates.}

\item{edges_to_keep}{A vector of directed arrows to be kept between (non-treatment and non-outcome) variables, e.g. c("Z1 -> Z2", "Z2 -> Z3"), c("y", "n", "y"), or c(TRUE, FALSE, TRUE).}
}
\value{
A dagitty object, with directed arrows removed based on edges_to_keep.
}
\description{
trimGraph() trims connected edges based on causal criteria and/or user inputs.
}
\examples{
#######################

# a fully connected graph
latent_variables <- c("Z3", "M1")
mediators <- c("M1", "M2")
confounders <- c("Z1", "Z2", "Z3")
treatment <- "X"
outcome <- "Y"
type <-"full"
dag <- buildGraph(type = type,
                                 treatment = treatment,
                                 outcome = outcome,
                                 confounders = confounders)

# get graph edges
edges_vec <- trimGraph(dag, edges_to_keep = FALSE) # enter 'n' to save edges

# Option 1: remove all edges between confounders
edges_to_keep <- c()
dag <- trimGraph(dag, edges_to_keep)
ggDagitty(dag, labels)

# Option 2: supply vector of kept edges
edges_to_keep <- c('Z1 -> Z2', 'Z3 -> Z2')
dag <- trimGraph(dag, edges_to_keep)
labels <- getLabels(dag)
ggDagitty(dag, labels)

# Option 3: run ESC-DAGs protocol and assess each edge with causal criteria
dag <- trimGraph(dag)
labels <- getLabels(dag)
ggDagitty(dag, labels)

# Option 4: keep edges 'y' or 'n'
edges_to_keep <- c("n", "n", "y", "n", "n", "y")
dag <- trimGraph(dag, edges_to_keep)
labels <- getLabels(dag)
ggDagitty(dag, labels)

# a saturated graph (confounders are not connected to mediators)
treatment <- "X"
outcome <- "Y"
confounders <- c("Z1", "Z2", "Z3")
mediators <- c("M1", "M2")
latent_variables <- c("Z3", "M1")
type <- "saturated"
coords_spec <- 3

dag <- causaliflower::buildGraph(type = type,
                                 treatment = treatment,
                                 outcome = outcome,
                                 confounders = confounders)

edges_to_keep <- c('Z1 -> Z2', 'Z3 -> Z2')
dag <- trimGraph(dag, edges_to_keep)

dag <- getCoords(dag,
                    confounders,
                    mediators = mediators,
                    coords_spec = coords_spec)

labels <- getLabels(dag)
ggDagitty(dag, labels)


# an ordered graph (confounders are connected in order of the supplied vector)
# adding coordinates and two mediator variables
confounders <- c("Z1", "Z2", "Z3")
treatment <- "X"
outcome <- "Y"
latents <- c("Z3", "M1") # latent confounder and mediator variables specified
mediators <- c("M1", "M2")
type <- "ordered"

dag <- buildGraph(type = type,
                 treatment = treatment,
                 outcome = outcome,
                 confounders = confounders,
                 mediators = mediators,
                 latent_variables = latent_variables)

dag <- trimGraph(dag)
labels <- getLabels(dag)
ggDagitty(dag, labels)

}
