% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{unobserved}
\alias{unobserved}
\title{Unobserved (latent) variables in a dagitty object}
\usage{
unobserved(dag)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
Character vector of unobserved (latent) names; character(0) if none.
}
\description{
unobserved() returns a vector of node names assigned "\code{[latent]}" in a dagitty object (similar to dagitty::latents())
}
\examples{
set.seed(1)
dag <- build_graph(
  variables = c("Z1", "Z2"), treatments = "X", outcomes = "Y",
  mediators = "M", instrumental_variables = "IV", type = "saturated"
)
unobserved(dag)

}
