% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_sensing_funcs.R
\name{validate_data_fusion}
\alias{validate_data_fusion}
\title{Validate MODIS-Landsat fusion accuracy via leave-one-out holdout}
\usage{
validate_data_fusion(
  longitude,
  latitude,
  start_date,
  end_date,
  fetch_metres = 600,
  max_cloud_cover = 50
)
}
\description{
Validate MODIS-Landsat fusion accuracy via leave-one-out holdout
}
