% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_utility_funcs.r
\name{variables}
\alias{variables}
\title{Variable names in a dagitty object}
\usage{
variables(dag, all_info = TRUE)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
A named vector of nodes in a dag.
}
\description{
variables() returns a named vector of variables corresponding to nodes in a dag, providing an easy way to extract node roles e.g. "treatment", "outcome", "confounder", "mediator", "latent", "mediator-outcome-confounder", or "instrumental".
}
