% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_coords_funcs.r
\name{add_coords}
\alias{add_coords}
\title{Generate new dag coordinates}
\usage{
add_coords(dag, coords_spec = 0, confounders = NULL)
}
\arguments{
\item{dag}{Dagitty object, with at least a treatment and outcome set. Nodes are automatically placed in the following categories: treatment, outcome, confounder, mediator, latent variable, mediator-outcome-confounder, or instrumental variable.}

\item{coords_spec}{Adjusts node placement, a higher value increases volatility and results in more extreme DAG structures.}

\item{confounders}{Vector of confounders in order of occurrence.}
}
\value{
dagitty object with coordinates.
}
\description{
Generate new dag coordinates
}
\examples{
dag <- add_coords(dag, coords_spec = 3) # update dagitty object node coordinates

plot_dagitty(dag) # check coordinates

}
