% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph.R
\name{add_nodes}
\alias{add_nodes}
\title{Add nodes to a dagitty object}
\usage{
add_nodes(
  dag,
  new_nodes,
  node_role = c("confounder", "treatment", "outcome", "mediator",
    "mediator_outcome_confounder", "instrumental", "competing_exposure", "collider",
    "latent", "observed"),
  type = c("full", "saturated", "first", "last"),
  temporal_reference_node = NA,
  print_edges = TRUE,
  coords_spec = 1
)
}
\arguments{
\item{dag}{A dagitty object. Must include exposure and outcome nodes.}

\item{new_nodes}{A suffix added to each of the new node names, e.g. "post_treatment", or "t" (a number is added for each repeat if num_repeats is specified)}

\item{node_role}{Role assigned to new nodes, from any of the following: c("confounder", "treatment", "outcome", "mediator", "mediator_outcome_confounder", "instrumental", "competing_exposure", "collider", "latent", "observed").}

\item{type}{Type of graph generated. Defaults to 'full' (fully connected graph) with arrows drawn between confounders (both directions) and from confounders to mediators. If type ='saturated', a similar saturated graph is produced except confounders are not connected to mediators, featuring bi-directional arrows between each of the confounders (follows the ESC-DAGs Mapping Stage in Ferguson et al. (2020)). When type = 'first' or 'last', inputted new_nodes are ordered first or last if a confounder, mediator, or treatment node_role is selected.}

\item{temporal_reference_node}{Supply an alternative reference, or simply leave blank. Default settings uses dagitty::topologicalOrdering() and selects the first of the inputted node_role (e.g., first confounder) as the temporal point of reference. If type = 'last', the last node is used.}

\item{coords_spec}{Set of parameters for generating coordinates. Adjust node placement with lambda, a higher value increases volatility and results in more extreme DAG structures.}
}
\value{
A dagitty object.
}
\description{
Add nodes to a dagitty object, connecting edges based on the 'type' of graph selected, and generate new node coordinates using existing nodes.
}
\examples{
dag <- add_nodes(dag,
                 new_nodes = new_node_vec,,
                 node_role = "confounder",
                 type = "last")

}
