% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph.R
\name{addNodes}
\alias{addNodes}
\title{Add nodes to a dagitty object}
\usage{
addNodes(
  dag,
  reference_nodes,
  reference_type = c("pre_treatment"),
  new_node_type = c("post_treatment"),
  num_repeats = NA
)
}
\arguments{
\item{dag}{A dagitty object. Must include exposure and outcome nodes.}

\item{reference_nodes}{Vector of existing node names, used as a reference for the new graph nodes, e.g., c("Z1", "Z2", "Z3").}

\item{reference_type}{A suffix added to each of the reference node names, e.g. "pre_treatment", or "t0".}

\item{new_node_type}{A suffix added to each of the new node names, e.g. "post_treatment", or "t" (a number is added for each repeat if num_repeats is specified)}

\item{num_repeats}{Number of additional copies of nodes, such as time points. Each repeat number is included at the end of new node names (new_new_t1, new_node_t2, etc.).}
}
\value{
A dagitty object.
}
\description{
addNodes() allows nodes to be added sequentially to a dagitty object using existing nodes as a reference point.
}
