% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_funcs.R
\name{assess_edges}
\alias{assess_edges}
\title{Assess dagitty object edges}
\usage{
assess_edges(dag, edges_to_keep = NA, assess_causal_criteria = FALSE)
}
\arguments{
\item{dag}{A dagitty object. Must include exposure and outcome nodes.}

\item{edges_to_keep}{A vector of directed arrows to be kept between (non-treatment and non-outcome) variables, e.g. c("Z1 -> Z2", "Z2 -> Z3"), c("y", "n", "y"), or c(TRUE, FALSE, TRUE).}

\item{assess_causal_criteria}{Defaults to FALSE. If TRUE, the user is guided through a sequence that assesses each pair of connected nodes using causal criteria. Based on the Evidence Synthesis for Constructing Directed Acyclic Graphs (ESC-DAGs) from Ferguson et al. (2020).}
}
\value{
A list or vector of edges.
}
\description{
assess_edges() provides ways to assess connected edges based on causal criteria and/or user inputs.
}
\examples{
## initial dag
{
  variables <- c("Z1", "Z2", "Z3") # these are treated as confounders
  treatments <- "X"
  outcomes <- "Y"
  type <- "ordered"

  dag <- build_graph(type = type,
                     variables = variables,
                     treatments = treatments,
                     outcomes = outcomes)
}

## Using a saturated graph as an example, created from an existing dag
saturated_graph <- build_graph(type = c('full', 'saturated'), # choose a type
                               variables = dag) # existing dagitty object inputted

## Option 1: existing dag as edges to keep
edges <- assess_edges(saturated_graph, edges_to_keep = dag)

## Option 2: guided causal criteria sequence (rule out edges already included in existing dag)
edges_to_keep <- assess_edges(saturated_graph, edges_to_keep = dag,
                              assess_causal_criteria = TRUE)

}
