% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_utility_funcs.r
\name{colliders}
\alias{colliders}
\title{Collider names in a dagitty object}
\usage{
colliders(dag, get_edges = FALSE)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
A vector of competing exposure names, or edges in a data table.
}
\description{
colliders() is a dagitty wrapper that identifies nodes in a directed acyclic graph connected to outcome, other than indicated exposures.
}
