% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_utility_funcs.r
\name{confounders}
\alias{confounders}
\title{Confounder names in a dagitty object}
\usage{
confounders(dag, get_edges = FALSE)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
A vector of confounder names, or edges in a data table.
}
\description{
confounders() is a dagitty wrapper that identifies common causes of treatment and outcome in a directed acyclic graph.
}
