% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funcs.r
\name{get_ancestor_edges}
\alias{get_ancestor_edges}
\title{Extract ancestor node edges}
\usage{
get_ancestor_edges(dag)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
Named list of edges.
}
\description{
ancestor_edges() returns a list of edges grouped by node names for an inputted dagitty object.
}
