% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funcs.r
\name{get_diff_edges}
\alias{get_diff_edges}
\title{dagitty node names, ancestor/descendant roles}
\usage{
get_diff_edges(dag1, dag2, compare_all = TRUE, include_roles = FALSE)
}
\arguments{
\item{dag1}{A dagitty object.}

\item{dag2}{A dagitty object.}

\item{compare_all}{Return all different edges (compare_all = TRUE) or only a subset, e.g. dag1 edges missing in dag2 (compare_all = FAlSE). Defaults to compare_all = TRUE.k}

\item{include_roles}{Determines whether to include edges with different node roles. Defaults to include_roles = FALSE, e.g. returns different irrespective of causal structure.}
}
\value{
Data table of edges.
}
\description{
node_structure() is a dagitty wrapper function that returns a list of node names extrated from a dagitty object, including the name and role of their ancestor and descendant nodes.
}
\examples{
get_diff(dag)

}
