% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funcs.r
\name{get_edges}
\alias{get_edges}
\title{dagitty node edges and roles}
\usage{
get_edges(
  dag,
  selected_nodes = c("outcome", "treatment", "confounder", "mediator",
    "mediator_outcome_confounder", "instrumental", "proxy", "competing_exposure",
    "collider", "latent", "observed")
)
}
\arguments{
\item{dag}{A dagitty object.}

\item{selected_nodes}{Nodes to return edges. Defaults to NULL, or can be a character or vector combination of any of the following: c("outcome", "treatment", "confounder", "mediator", "mediator_outcome_confounder", "instrumental", "proxy", "competing_exposure", "collider", "latent", "observed"))}
}
\value{
A data frame, data table, or list of edges for the roles specified in selected_nodes.
}
\description{
get_edges() filters a dagitty object and returns a data frame with edges for specified node roles.
}
\examples{
edges <- get_edges(dag)

}
