% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funcs.r
\name{get_labels}
\alias{get_labels}
\title{Label DAG nodes
Extract node initials to use as labels in plot_dagitty, or dag plots using ggdag and ggplot2. This is essentially a wrapper function for ggdag::dag_label().}
\usage{
get_labels(dag)
}
\arguments{
\item{dag}{A dagittyy object.}
}
\value{
Labelled vector of node names from the dagitty object
}
\description{
Label DAG nodes
Extract node initials to use as labels in plot_dagitty, or dag plots using ggdag and ggplot2. This is essentially a wrapper function for ggdag::dag_label().
}
