% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assess_keep_edges.r
\name{get_roles}
\alias{get_roles}
\title{dagitty nodes grouped by role}
\usage{
get_roles(dag, multiple_roles = FALSE)
}
\arguments{
\item{dag}{dagitty object}
}
\value{
Nested list of nodes and node relationships
}
\description{
dagitty nodes grouped by role
}
