% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funcs.r
\name{get_roles}
\alias{get_roles}
\title{dagitty nodes grouped by role}
\usage{
get_roles(dag, multiple_roles = FALSE)
}
\arguments{
\item{dag}{A dagitty object.}

\item{multiple_roles}{Defaults to FALSE (one role per node). If set to TRUE, multiple roles can be returned for some nodes (e.g. latent mediator variable).}
}
\value{
Nested list of nodes and node relationships
}
\description{
dagitty nodes grouped by role
}
\examples{
get_roles(dag)

}
