% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_funcs.r
\name{get_structure}
\alias{get_structure}
\title{dagitty node names, ancestor/descendant roles}
\usage{
get_structure(dag)
}
\arguments{
\item{dag}{dagitty object}
}
\value{
Nested list of nodes and node relationships
}
\description{
node_structure() is a dagitty wrapper function that returns a list of node names extrated from a dagitty object, including the name and role of their ancestor and descendant nodes.
}
\examples{
get_structure(dag)

}
