% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edges_feature_map.r
\name{getEdges}
\alias{getEdges}
\title{Gets edges between nodes and their roles}
\usage{
getEdges(
  dag,
  selected_nodes = c("outcome", "treatment", "confounder", "mediator",
    "mediator_outcome_confounder", "instrumental", "proxy", "competing_exposure",
    "collider", "latent"),
  variables = NULL,
  output_format = "default",
  output_structure = "data.table"
)
}
\arguments{
\item{dag}{A dagitty object.}

\item{selected_nodes}{Nodes to return edges. Defaults to NULL, or can be a character or vector combination of any of the following: c("treatment", "outcome", "confounder", "mediator", "latent", "mediator-outcome-confounder", "instrumental")}

\item{output_format}{Options for outputted data format: "default" returns a data frame with separate columns for ancestor and descendant roles, "wide" returns separate columns for each unique role for both ancestor and descendant nodes ("list" returns the same wide data format except as a list), and "long" assigns separate rows to ancestor and descendant nodes.}

\item{output_structure}{Outputted data can be a "data.table", "data.frame", or "list".}
}
\value{
A data frame, data table, or list of edges for the roles specified in selected_nodes.
}
\description{
getEdges() filters a dagitty object and returns a data frame with edges for specified node roles.
}
