% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edges_feature_map.r
\name{getFeatureMap}
\alias{getFeatureMap}
\title{Convert dagitty object to list}
\usage{
getFeatureMap(dag, include_descendant_nodes = FALSE)
}
\arguments{
\item{dag}{dagitty object}
}
\value{
Nested list of nodes and node relationships
}
\description{
Convert dagitty object to list
}
