% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assess_keep_edges.r
\name{getFeatureMap}
\alias{getFeatureMap}
\title{List of dag node names}
\usage{
getFeatureMap(dag)
}
\arguments{
\item{dag}{dagitty object}
}
\value{
Nested list of nodes and node relationships
}
\description{
List of dag node names
}
