% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdagitty.R
\name{getLabels}
\alias{getLabels}
\title{Label DAG nodes
Extract node initials to use as labels in ggdagitty, or dag plots using ggdag and ggplot2. This is essentially a wrapper function for ggdag::dag_label().}
\usage{
getLabels(dag)
}
\arguments{
\item{dag}{A dagittyy object.}
}
\value{
Labelled vector of node names from the dagitty object
}
\description{
Label DAG nodes
Extract node initials to use as labels in ggdagitty, or dag plots using ggdag and ggplot2. This is essentially a wrapper function for ggdag::dag_label().
}
