% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{instrumental_variables}
\alias{instrumental_variables}
\title{Extracts instrumental variable names from a dagitty object}
\usage{
instrumental_variables(dag)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
Vector of instrumental variable names.
}
\description{
getInstrumentalVariables() is a dagitty wrapper function capable of identifying  multiple instrumental variables in multi-treatment and multi-outcome settings.
}
\examples{
instrumental_variables(dag)

}
