% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{instruments}
\alias{instruments}
\title{Get instrumental variable names}
\usage{
instruments(dag)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
List of instrumental variable names.
}
\description{
instruments() is a dagitty wrapper function capable of identifying  multiple instrumental variables in multi-treatment and multi-outcome settings.
}
\examples{
instruments(dag)

}
