% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_funcs.R
\name{join_graphs}
\alias{join_graphs}
\title{Merge two dagitty objects}
\usage{
join_graphs(dag, new_dag, coords_spec = 0.1, threshold = 0.5)
}
\arguments{
\item{dag}{First dagitty object.}

\item{new_dag}{Second dagitty object, added to the first.}

\item{coords_spec}{Adjust node placement when generating coordinates with a numeric value > 0 (recommended < 10).}

\item{threshold}{Set the allowed closeness of nodes in newly generated coordinates. Numeric value between 0 and 1 for the minimum distance allowed when generating coordinates.}
}
\value{
A dagitty object
}
\description{
join_graphs() adds new nodes to the first from the second supplied dagitty object, before generating new coordinates for the merged graph.
}
\examples{
new_graph <- join_graphs(dag, new_dag)

}
