% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_funcs.R
\name{keep_edges}
\alias{keep_edges}
\title{Remove dagitty object edges}
\usage{
keep_edges(dag, edges_to_keep = NA)
}
\arguments{
\item{dag}{A saturated graph dagitty object. Exposure and outcome must be indicated, and optionally can include assigned coordinates.}

\item{edges_to_keep}{A vector of directed arrows to be kept between (non-treatment and non-outcome) variables, e.g. c("Z1 -> Z2", "Z2 -> Z3"), c("y", "n", "y"), or c(TRUE, FALSE, TRUE).}
}
\value{
A dagitty object, with directed arrows removed based on edges_to_keep.
}
\description{
keep_edges() removes edges based on user inputs.
}
\examples{
## initial dag
{
  variables <- c("Z1", "Z2", "Z3") # these are treated as confounders
  treatments <- "X"
  outcomes <- "Y"
  type <- "ordered"

  dag <- build_graph(type = type,
                     variables = variables,
                     treatments = treatments,
                     outcomes = outcomes)
}

## Using a saturated graph as an example, created from an existing dag
saturated_graph <- build_graph(type = c('full', 'saturated'), # choose a type
                               variables = dag) # existing dagitty object inputted

## guided causal criteria sequence (rule out edges already included in existing dag)
edges_to_keep <- assess_edges(saturated_graph, edges_to_keep = dag,
                              assess_causal_criteria = TRUE)

dag <- keep_edges(saturated_graph, edges_to_keep)

}
