% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_sensing_funcs.R
\docType{data}
\name{landsat_index_list}
\alias{landsat_index_list}
\title{Landsat spectral indices list}
\format{
A named list where each element is a list with components:
\describe{
\item{assets}{Character vector of Sentinel-2 band names.}
\item{fun}{A function that takes a SpatRaster and returns a single-layer SpatRaster.}
\item{res}{Recommended resolution (highest integer value used by assets).}
}
}
\usage{
landsat_index_list
}
\description{
A named list of spectral indices compatible with Landsat 5 & 8 sources.
}
\examples{
names(landsat_index_list)

ndvi_df <- get_rs_data(longitude = 138.715,
                      latitude = -34.904,
                      x_metres = 30,
                      y_metres = 30,
                      start_date = "2023-01-01",
                      end_date = "2023-01-10",
                      index_name = "NDVI")

}
\keyword{datasets}
