% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_utility_funcs.r
\name{mediator_outcome_confounders}
\alias{mediator_outcome_confounders}
\title{Mediator-outcome confounder names in a dagitty object}
\usage{
mediator_outcome_confounders(dag, edges = FALSE)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
A vector of mediator-outcome confounder names, or edges in a data table.
}
\description{
moc() is a dagitty wrapper that identifies nodes along paths between treatment and outcome in a directed acyclic graph.
}
