% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDag.R
\name{mediators}
\alias{mediators}
\title{Mediator names in a dagitty object}
\usage{
mediators(dag, edges = FALSE)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
A vector of mediators, or dataframe of edges.
}
\description{
mediators() is a dagitty wrapper that identifies nodes along paths between treatment and outcome in a directed acyclic graph.
}
