% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_graph.R
\name{merge_graphs}
\alias{merge_graphs}
\title{Merge two dagitty objects}
\usage{
merge_graphs(dag, new_dag, temporal_reference_node = NA, coords_spec = 0.1)
}
\arguments{
\item{dag}{First dagitty object.}

\item{new_dag}{Second dagitty object, added to the first.}

\item{temporal_reference_node}{Supply an alternative reference, or simply leave blank. Default settings use treatment as the temporal point of reference for adding new (post-treatment) nodes, but if other node types are specified it can be useful to specify the existing node name.}

\item{coords_spec}{Set of parameters for generating coordinates. Adjust node placement with lambda, a higher value increases volatility and results in more extreme DAG structures.}
}
\value{
A dagitty object
}
\description{
merge_graphs() adds new nodes to the first from the second supplied dagitty object, before generating new coordinates for the merged graph.
}
\examples{
merged_dag <- merge_graphs(dag, new_dag)

}
