% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_sensing_funcs.R
\name{metres_to_degrees}
\alias{metres_to_degrees}
\title{Convert metres to decimal degrees}
\usage{
metres_to_degrees(latitude, distance = 1, pixel_size = 1)
}
\description{
Convert metres to decimal degrees
}
