% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_utility_funcs.r
\name{minimal_sets}
\alias{minimal_sets}
\title{Minimally sufficient adjustment sets}
\usage{
minimal_sets(
  dag,
  treatment = NULL,
  outcome = NULL,
  effect = "total",
  smallest = 5,
  decreasing = FALSE
)
}
\arguments{
\item{dag}{A dagitty object.}

\item{treatment}{Vector of treatment(s).}

\item{outcome}{Vector of outcome(s).}

\item{effect}{Defaults to total effect, options available as per dagitty::adjustmentSets()}

\item{smallest}{Number of sets to return, defaults to the smallest five minimally sufficient adjustment sets.}

\item{decreasing}{Defaults to FALSE (showing smallest). Optionally can be set to filter the largest minimally sufficient adjustment sets.}
}
\value{
Named list of minimally sufficient adjustment sets.
}
\description{
minimal_sets() is a dagitty::adjustmentSets() wrapper for obtaining minimal adjustment sets, returning the smallest 5 (if available) by default.
}
