% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalDag.R
\name{moc}
\alias{moc}
\title{Mediator-outcome confounder names in a dagitty object}
\usage{
moc(dag, edges = FALSE)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
A vector of mediator-outcome confounders, or dataframe of edges.
}
\description{
moc() is a dagitty wrapper that identifies nodes along paths between treatment and outcome in a directed acyclic graph.
}
