% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_and_coords_funcs.r
\name{plot_dagitty}
\alias{plot_dagitty}
\title{ggdag plot from a dagitty object}
\usage{
plot_dagitty(
  dag,
  coords_spec = 0,
  labels = NULL,
  label_type = "name",
  label_placement = "label_repel"
)
}
\arguments{
\item{dag}{A dagitty object}

\item{coords_spec}{Adjusts node placement, a higher value increases volatility and results in more extreme DAG structures.}

\item{labels}{A vector of labels for nodes in dagitty object}

\item{label_type}{Defaults to using the node names of the inputted dag, or 'initials' if specified. Labels are only assigned by plot_dagitty() if nodes are not already labelled, using a 'ggdag::dag_label()' wrapper..}

\item{label_placement}{Labels are placed in a text box adjacent each node by default, or can be placed inside each node using 'label_node'. It is recommended to also specify label_type = 'initials' where label_placement = 'label_node' is used.}
}
\value{
ggdag plot
}
\description{
Generates a ggdag plot from a dagitty object, adding coordinates and node labels using initials (capital letters).
Essentially a wrapper for dagitty and ggdag plotting functions with added features.
}
\examples{
plot_dagitty(dag)

plot_dagitty(dag, coords_spec = 3) # increasing spec for more variation in node placement

}
