% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_sensing_funcs.R
\name{point_to_bbox}
\alias{point_to_bbox}
\title{Bounding box around a central point}
\usage{
point_to_bbox(longitude, latitude, x_metres, y_metres)
}
\arguments{
\item{longitude}{Longitude in decimal degrees (WGS84).}

\item{latitude}{Latitude in decimal degrees (WGS84).}

\item{x_metres}{Bbox x width in metres.}

\item{y_metres}{Bbox y length in metres.}
}
\value{
An sf bbox object.
}
\description{
Bounding box around a central point
}
