% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{proxies}
\alias{proxies}
\title{Proxy node names in a dagitty object}
\usage{
proxies(dag)
}
\arguments{
\item{dag}{A dagitty object.}
}
\value{
A vector of proxy variable names, or edges in a data table.
}
\description{
proxies() is a dagitty wrapper that identifies nodes that are proxy variables for indicated unmeasured confounding (existing along a path between a latent variable and outcome) in a directed acyclic graph.
}
\examples{
proxies(dag)

}
